refactor: isolate database statistics reporting
